refactor(pty): extract reversible-stop ownership from kill.ts
Rebasing onto main pushed kill.ts past the 300-line max-lines budget; a disable or ratchet bump is not allowed, so the ref-counted reversible-stop bookkeeping moves to its own module.
