+++ src/main/ipc/pty/runtime/reversible-stop-ownership.ts
@@ -1,0 +1,27 @@
+import type { PtyRuntimeControllerDeps } from './controller-deps'
+
+/** Ref-counted so nested reversible-stop scopes cannot clear a sibling's mark on the first release. */
+export function markReversibleStopsFromRuntimeController(
+  deps: PtyRuntimeControllerDeps,
+  ptyIds: readonly string[]
+): () => void {
+  const { reversibleStopOwnersByPtyId } = deps
+  for (const ptyId of ptyIds) {
+    reversibleStopOwnersByPtyId.set(ptyId, (reversibleStopOwnersByPtyId.get(ptyId) ?? 0) + 1)
+  }
+  let released = false
+  return () => {
+    if (released) {
+      return
+    }
+    released = true
+    for (const ptyId of ptyIds) {
+      const owners = (reversibleStopOwnersByPtyId.get(ptyId) ?? 0) - 1
+      if (owners > 0) {
+        reversibleStopOwnersByPtyId.set(ptyId, owners)
+      } else {
+        reversibleStopOwnersByPtyId.delete(ptyId)
+      }
+    }
+  }
+}
